src/DeepCard.API/requirements.txt
asn1crypto==0.24.0
bleach==1.5.0
cryptography==2.1.4
cycler==0.10.0
Cython==0.24
easydict==1.9
h5py==2.9.0
html5lib==0.9999999
idna==2.6
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
protobuf==3.7.1
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.0
python-dateutil==2.8.0
pyxdg==0.25
PyYAML==5.1
scipy==1.3.0
SecretStorage==2.3.1
six==1.11.0
torch==1.1.0
torchvision==0.3.0
Werkzeug==0.15.4
